requirements.txt
discord.py>=2.3.0
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
flask>=3.0.0
PyNaCl>=1.5.0

